updated transformers, pipelines & machine learning data creation
