+++ dict.json
@@ -1,0 +1,1 @@
+{"normal": 0, "fatigue": 1, "awake": 2, "unlabeled": 3}
